+++ core/src/zeit/retresco/tests/test_content.py
@@ -1,12 +1,12 @@
 from zeit.cms.checkout.helper import checked_out
 from zeit.content.article.interfaces import IArticle
 import zeit.cms.content.sources
+import zeit.cms.tagging.tag
 import zeit.content.article.article
 import zeit.content.author.author
 import zeit.content.image.interfaces
 import zeit.content.link.interfaces
 import zeit.content.volume.volume
-import zeit.retresco.tag
 import zeit.retresco.testing
 import zope.schema
 
@@ -40,7 +40,7 @@ class ContentTest(zeit.retresco.testing.FunctionalTestCase):
         author.lastname = u'Shakespeare'
         self.repository['shake'] = author
         with checked_out(article) as co:
-            co.keywords = (zeit.retresco.tag.Tag('Berlin', 'location'),)
+            co.keywords = (zeit.cms.tagging.tag.Tag('Berlin', 'location'),)
             co.authorships = (co.authorships.create(self.repository['shake']),)
             co.agencies = (self.repository['shake'],)
         article = zeit.cms.interfaces.ICMSContent(
