+++ core/src/zeit/retresco/tests/test_tagger.py
@@ -2,7 +2,7 @@
 from unittest import mock
 from zeit.cms.checkout.helper import checked_out
 from zeit.cms.testcontenttype.testcontenttype import ExampleContentType
-from zeit.retresco.tag import Tag
+from zeit.cms.tagging.tag import Tag
 from zeit.retresco.tagger import Tagger
 from zeit.retresco.testing import create_testcontent
 import lxml.objectify
